+++ cs-rft/src/main/java/io/cloudslang/content/rft/remote_copy/CopierFactory.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudslang.content.rft.spike_rft;
+package io.cloudslang.content.rft.remote_copy;
 
 import java.util.HashMap;
 
@@ -24,10 +24,7 @@ public class CopierFactory {
     static {
         executorMap.put(copiers.scp, ScpCopier.class);
         executorMap.put(copiers.local, LocalCopier.class);
-        executorMap.put(copiers.ftp, FtpCopier.class);
         executorMap.put(copiers.sftp, SftpCopier.class);
-        executorMap.put(copiers.smb, SmbCopier.class);
-        executorMap.put(copiers.smb2, SmbCopier.class);
         executorMap.put(copiers.smb3, SmbCopier.class);
     }
 
@@ -43,5 +40,5 @@ public class CopierFactory {
         return executorMap.get(name).newInstance();
     }
 
-    public enum copiers {local, scp, ftp, sftp, smb, smb2, smb3}
+    public enum copiers {local, scp, sftp, smb3}
 }
