+++ cs-rft/src/main/java/io/cloudslang/content/rft/remote_copy/ICopier.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudslang.content.rft.spike_rft;
+package io.cloudslang.content.rft.remote_copy;
 
 public interface ICopier {
