+++ cs-rft/src/main/java/io/cloudslang/content/rft/remote_copy/SimpleReader.java
@@ -12,13 +12,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cloudslang.content.rft.spike_rft;
+package io.cloudslang.content.rft.remote_copy;
 
 import java.io.File;
 
-public interface IReader {
+public class SimpleReader implements IReader {
+    String name;
+    File file;
 
-    public String getFileName();
+    SimpleReader(String fileName, File file) {
+        name = fileName;
+        this.file = file;
+    }
 
-    public File getFile();
+    public String getFileName() {
+        return name;
+    }
+
+    public File getFile() {
+        return file;
+    }
 }
