+++ cs-rft/src/main/java/io/cloudslang/content/rft/utils/InputsValidation.java
@@ -175,5 +175,4 @@ public class InputsValidation {
         return exceptions;
     }
 
-
 }
